requirements.txt
PyQt6>=6.5.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
openpyxl>=3.1.0
pandas>=2.0.0
Pillow>=10.0.0
regex>=2023.6.3

# Optional: For creating executables
pyinstaller>=5.0.0
auto-py-to-exe>=2.0.0
